Remove dedicated crate
