+++ rust/connlib/clients/android/Cargo.toml
@@ -16,7 +16,6 @@ tracing-android = "0.2"
 tracing = { version = "0.1", features = ["std", "attributes"] }
 tracing-subscriber = "0.3"
 tracing-appender = "0.2"
-tracing-on-demand-rolling-appender = { workspace = true }
 firezone-client-connlib = { path = "../../libs/client" }
 jni = { version = "0.21.1", features = ["invocation"] }
 ip_network = "0.4"
