+++ rust/connlib/clients/apple/Cargo.toml
@@ -23,7 +23,6 @@ tracing = "0.1"
 tracing-oslog = { git = "https://github.com/sbag13/tracing-oslog", rev = "0f82b8051c65de86191e1350afc7a26d5c670c29" }
 tracing-subscriber = "0.3"
 tracing-appender = "0.2"
-tracing-on-demand-rolling-appender = { workspace = true }
 
 [lib]
 name = "connlib"
