+++ rust/connlib/libs/client/Cargo.toml
@@ -19,7 +19,6 @@ serde = { version = "1.0", default-features = false, features = ["std", "derive"
 boringtun = { workspace = true }
 backoff = { workspace = true }
 url = { version = "2.4.1", features = ["serde"] }
-tracing-on-demand-rolling-appender = { workspace = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 tracing = { version = "0.1", default-features = false, features = ["std", "attributes"] }
