+++ rust/connlib/libs/client/src/file_logger.rs
@@ -13,7 +13,9 @@
 //! - Device serials
 //! - MAC addresses
 
-use std::path::Path;
+use std::path::{Path, PathBuf};
+use std::sync::{Arc, Mutex};
+use std::{fs, io};
 use tracing::{level_filters::LevelFilter, Subscriber};
 use tracing_subscriber::Layer;
 
@@ -25,7 +27,7 @@ pub fn layer<T>(
 ) -> (
     Box<dyn Layer<T> + Send + Sync + 'static>,
     tracing_appender::non_blocking::WorkerGuard,
-    tracing_on_demand_rolling_appender::Handle,
+    Handle,
 )
 where
     T: Subscriber + for<'a> tracing_subscriber::registry::LookupSpan<'a>,
@@ -36,10 +38,7 @@ where
     #[cfg(not(debug_assertions))]
     let level = LevelFilter::WARN;
 
-    let (appender, handle) = tracing_on_demand_rolling_appender::new(
-        log_dir.to_path_buf(),
-        LOG_FILE_BASE_NAME.to_owned(),
-    );
+    let (appender, handle) = new_appender(log_dir.to_path_buf(), LOG_FILE_BASE_NAME.to_owned());
 
     let (writer, guard) = tracing_appender::non_blocking(appender);
 
@@ -52,4 +51,66 @@ where
     // we have to wait for tracing_appender to flush the logs before exiting.
     // See https://docs.rs/tracing-appender/latest/tracing_appender/non_blocking/struct.WorkerGuard.html
     (layer, guard, handle)
+}
+
+fn new_appender(directory: PathBuf, name: String) -> (Appender, Handle) {
+    let inner = Arc::new(Mutex::new(Inner { directory, name }));
+    let appender = Appender {
+        inner: inner.clone(),
+    };
+    let handle = Handle { inner };
+
+    (appender, handle)
+}
+
+#[derive(Clone, Debug)]
+pub struct Handle {
+    inner: Arc<Mutex<Inner>>,
+}
+
+#[derive(Debug)]
+struct Appender {
+    inner: Arc<Mutex<Inner>>,
+}
+
+impl Handle {
+    /// Rolls over to a new file.
+    ///
+    /// Returns the path to the now unused, previous log file.
+    pub fn roll_to_new_file(&self) -> PathBuf {
+        todo!()
+    }
+}
+
+#[derive(Debug)]
+struct Inner {
+    directory: PathBuf,
+    name: String,
+}
+
+impl io::Write for Appender {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        todo!()
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        todo!()
+    }
+}
+
+// Copied from `tracing-appender/src/rolling.rs`.
+fn create_writer(directory: &str, filename: &str) -> io::Result<fs::File> {
+    let path = Path::new(directory).join(filename);
+    let mut open_options = fs::OpenOptions::new();
+    open_options.append(true).create(true);
+
+    let new_file = open_options.open(path.as_path());
+    if new_file.is_err() {
+        if let Some(parent) = path.parent() {
+            fs::create_dir_all(parent)?;
+            return open_options.open(path);
+        }
+    }
+
+    new_file
 }
