--- rust/tracing-on-demand-rolling-appender/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "tracing-on-demand-rolling-appender"
-version = "0.1.0"
-edition = "2021"
